feat: add scroll fix when type and fix bugs
when scrolling to bottom and then typing, scrolling resets to 0 if
list doesn't fill up display (only scrolling when many elements)
also fix issue that when type and remove and type same letter not
registered because it used to only update input when typing and not
when input empty. Also refactoring a bit.